requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
anthropic==0.40.0
gspread==6.1.2
google-auth==2.35.0
